+++ app/models/peer_review.rb
@@ -84,6 +84,12 @@ class PeerReview < ApplicationRecord
     self.where(reviewer_id: reviewer_group).size
   end
 
+  def self.get_num_collected(reviewer_group)
+    Grouping.joins(:peer_reviews)
+            .where('peer_reviews.reviewer_id': reviewer_group)
+            .where(is_collected: true).count
+  end
+
   def self.get_num_marked(reviewer_group)
     self.includes(:result).where(reviewer_id: reviewer_group).count do |pr|
       pr.result.marking_state == Result::MARKING_STATES[:complete]