+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## [v1.10.2]
 - Fixed the Assign Reviewers table loading issue (#4894)
+- Fixed a bug where the progress bar in submissions and results page counts the not collected submissions (#4854)
 
 ## [v1.10.1]
 - Fix out of dates link to the wiki (#4843)